Validate arguments in value object
JS problems, for any reason, if i send a request to create a new book without name or length, it will be created without any error.
There is a very similar problem if you send, for example, 7 instead a string name as a book atribute. It will be created anyway